Travail sur le sélecteur de grille
