+++ src/AfficheurSelection.rb
@@ -1,8 +1,10 @@
 require "gtk3"
-require_relative "SelectionGrille"
+require_relative "SelecteurGrille"
 
+# Widget Gtk permettant d'afficher des sélectionneurs de grille
 class AfficheurSelection < Gtk::Box
 
+	# Constructeur
 	def initialize()
 		super(Gtk::Orientation.new(0), 0)
 
@@ -17,7 +19,7 @@ class AfficheurSelection < Gtk::Box
 		boxHorizontale.add(stackDif)
 
 		["Facile", "Moyen", "Difficile"].each() do |dif|
-			stackDif.add_titled(SelectionGrille.new(), dif, dif)
+			stackDif.add_titled(SelecteurGrille.new(), dif, dif)
 		end
 	end
 end
